Programm without input checks
